requirements.txt
Pillow==7.2.0
PyMuPDF==1.17.5
PyPDF2==1.26.0
pytesseract==0.3.5
